+++ packages/sdks-tests/src/e2e-tests/editing.spec.ts
@@ -3,14 +3,14 @@ import { MODIFIED_COLUMNS } from '../specs/columns.js';
 import { MODIFIED_EDITING_STYLES } from '../specs/editing-styles.js';
 import { NEW_TEXT } from '../specs/helpers.js';
 import { MODIFIED_HOMEPAGE } from '../specs/homepage.js';
-import { test } from '../helpers/index.js';
+import { checkIsRN, test } from '../helpers/index.js';
 import { launchEmbedderAndWaitForSdk, sendContentUpdateMessage } from '../helpers/visual-editor.js';
+import { MODIFIED_EDITING_COLUMNS } from '../specs/editing-columns-inner-layout.js';
 
 const editorTests = ({ noTrustedHosts }: { noTrustedHosts: boolean }) => {
   test('correctly updates Text block', async ({ page, basePort, packageName }) => {
     test.skip(
-      packageName === 'react-native' ||
-        packageName === 'nextjs-sdk-next-app' ||
+      packageName === 'nextjs-sdk-next-app' ||
         packageName === 'gen1-next' ||
         packageName === 'gen1-react' ||
         packageName === 'gen1-remix'
@@ -27,8 +27,7 @@ const editorTests = ({ noTrustedHosts }: { noTrustedHosts: boolean }) => {
 
   test('correctly updates Text block styles', async ({ page, packageName, basePort }) => {
     test.skip(
-      packageName === 'react-native' ||
-        packageName === 'nextjs-sdk-next-app' ||
+      packageName === 'nextjs-sdk-next-app' ||
         packageName === 'gen1-next' ||
         packageName === 'gen1-react' ||
         packageName === 'gen1-remix'
@@ -56,8 +55,7 @@ test.describe('Visual Editing', () => {
     packageName,
   }) => {
     test.skip(
-      packageName === 'react-native' ||
-        packageName === 'nextjs-sdk-next-app' ||
+      packageName === 'nextjs-sdk-next-app' ||
         packageName === 'gen1-next' ||
         packageName === 'gen1-react' ||
         packageName === 'gen1-remix'
@@ -67,6 +65,57 @@ test.describe('Visual Editing', () => {
     await sendContentUpdateMessage({ page, newContent: MODIFIED_COLUMNS, model: 'page' });
     await page.frameLocator('iframe').getByText(NEW_TEXT).waitFor();
   });
+
+  test('correctly updates Box -> Columns when used Inner Layout > Columns option', async ({
+    page,
+    packageName,
+    basePort,
+    sdk,
+  }) => {
+    test.skip(
+      packageName === 'nextjs-sdk-next-app' ||
+        packageName === 'gen1-next' ||
+        packageName === 'gen1-react' ||
+        packageName === 'gen1-remix'
+    );
+
+    await launchEmbedderAndWaitForSdk({
+      path: '/editing-box-columns-inner-layout',
+      basePort,
+      page,
+    });
+
+    const firstText = page.frameLocator('iframe').getByText('first');
+    const secondText = page.frameLocator('iframe').getByText('second');
+    await expect(firstText).toBeVisible();
+    await expect(secondText).toBeVisible();
+    const firstBox = await firstText.boundingBox();
+    const secondBox = await secondText.boundingBox();
+    if (firstBox && secondBox) {
+      expect(firstBox.y).toBeLessThan(secondBox.y);
+    }
+
+    await sendContentUpdateMessage({ page, newContent: MODIFIED_EDITING_COLUMNS, model: 'page' });
+    // had to hack this so that we can wait for the content update to actually show up (was failing in Qwik)
+    await page.frameLocator('iframe').getByText('third').waitFor();
+
+    const updatedFirstText = page.frameLocator('iframe').getByText('third');
+    await expect(updatedFirstText).toBeVisible();
+    await expect(secondText).toBeVisible();
+    const updatedFirstBox = await updatedFirstText.boundingBox();
+    const updatedSecondBox = await secondText.boundingBox();
+    if (updatedFirstBox && updatedSecondBox) {
+      if (checkIsRN(sdk)) {
+        // stack layout incase of RN SDK
+        expect(updatedFirstBox.x).toBe(updatedSecondBox.x);
+        expect(updatedFirstBox.y).toBeLessThan(updatedSecondBox.y);
+      } else {
+        expect(updatedFirstBox.x).toBeLessThan(updatedSecondBox.x);
+        expect(updatedFirstBox.y).toBe(updatedSecondBox.y);
+      }
+    }
+  });
+
   test.describe('fails for empty trusted hosts', () => {
     test.fail();
     editorTests({ noTrustedHosts: true });