+++ packages/sdks-tests/src/e2e-tests/external-data.spec.ts
@@ -1,16 +1,14 @@
 import { expect } from '@playwright/test';
-import { excludeTestFor, test } from '../helpers/index.js';
+import { test } from '../helpers/index.js';
 
 test.describe('External Data', () => {
-  test('renders initial value', async ({ page, packageName, sdk }) => {
-    test.fail(excludeTestFor({ angular: true }, sdk), 'Angular Gen2 SDK not implemented.');
+  test('renders initial value', async ({ page, packageName }) => {
     test.skip(packageName !== 'react');
     await page.goto('/external-data');
 
     await expect(page.getByText('Data value: foo')).toBeVisible();
   });
-  test('rerenders on external change', async ({ page, packageName, sdk }) => {
-    test.fail(excludeTestFor({ angular: true }, sdk), 'Angular Gen2 SDK not implemented.');
+  test('rerenders on external change', async ({ page, packageName }) => {
     test.skip(packageName !== 'react');
     await page.goto('/external-data');