+++ packages/sdks-tests/src/specs/editing-columns-inner-layout.ts
@@ -1,0 +1,198 @@
+export const EDITING_BOX_TO_COLUMN_INNER_LAYOUT = {
+  ownerId: 'ad30f9a246614faaa6a03374f83554c9',
+  lastUpdateBy: null,
+  createdDate: 1724132675778,
+  id: 'fd46ebf82e0c481681deb9e0823faeca',
+  '@version': 4,
+  name: 'asfxgxgv',
+  modelId: '17c6065109ef4062ba083f5741f4ee6a',
+  published: 'draft',
+  meta: {
+    hasLinks: false,
+    kind: 'page',
+    lastPreviewUrl:
+      'http://localhost:5173/asfxgxgv?builder.space=ad30f9a246614faaa6a03374f83554c9&builder.user.permissions=read%2Ccreate%2Cpublish%2CeditCode%2CeditDesigns%2Cadmin%2CeditLayouts%2CeditLayers&builder.user.role.name=Admin&builder.user.role.id=admin&builder.cachebust=true&builder.preview=page&builder.noCache=true&builder.allowTextEdit=true&__builder_editing__=true&builder.overrides.page=fd46ebf82e0c481681deb9e0823faeca&builder.overrides.fd46ebf82e0c481681deb9e0823faeca=fd46ebf82e0c481681deb9e0823faeca&builder.overrides.page:/asfxgxgv=fd46ebf82e0c481681deb9e0823faeca&builder.options.locale=Default',
+  },
+  priority: -611,
+  query: [
+    {
+      '@type': '@builder.io/core:Query',
+      property: 'urlPath',
+      operator: 'is',
+      value: '/asfxgxgv',
+    },
+  ],
+  data: {
+    themeId: false,
+    title: 'asfxgxgv',
+    inputs: [],
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-caf4d18da3d0439281f29eb144874e2e',
+        children: [
+          {
+            '@type': '@builder.io/sdk:Element',
+            '@version': 2,
+            id: 'builder-6a7ca44506bf459abaff59e2e76fdaaf',
+            component: {
+              name: 'Text',
+              options: {
+                text: 'first',
+              },
+            },
+            responsiveStyles: {
+              large: {
+                display: 'flex',
+                flexDirection: 'column',
+                position: 'relative',
+                flexShrink: '0',
+                boxSizing: 'border-box',
+                marginTop: '20px',
+                lineHeight: 'normal',
+                height: 'auto',
+              },
+            },
+          },
+          {
+            '@type': '@builder.io/sdk:Element',
+            '@version': 2,
+            id: 'builder-f102f51c56b84727be3f222c1d34b3ea',
+            component: {
+              name: 'Text',
+              options: {
+                text: 'second',
+              },
+            },
+            responsiveStyles: {
+              large: {
+                display: 'flex',
+                flexDirection: 'column',
+                position: 'relative',
+                flexShrink: '0',
+                boxSizing: 'border-box',
+                marginTop: '20px',
+                lineHeight: 'normal',
+                height: 'auto',
+              },
+            },
+          },
+        ],
+        responsiveStyles: {
+          large: {
+            display: 'flex',
+            flexDirection: 'column',
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            height: 'auto',
+            paddingBottom: '30px',
+          },
+        },
+      },
+    ],
+  },
+  metrics: {
+    clicks: 0,
+    impressions: 0,
+  },
+  variations: {},
+  lastUpdated: 1724165281228,
+  testRatio: 1,
+  createdBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+  lastUpdatedBy: 'RuGeCLr9ryVt1xRazFYc72uWwIK2',
+  folders: [],
+};
+
+export const MODIFIED_EDITING_COLUMNS = {
+  id: 'fd46ebf82e0c481681deb9e0823faeca',
+  data: {
+    themeId: false,
+    title: 'asfxgxgv',
+    inputs: [],
+    blocks: [
+      {
+        '@type': '@builder.io/sdk:Element',
+        '@version': 2,
+        id: 'builder-a5309f6a1d2c43068e51557edcc0a85d',
+        component: {
+          name: 'Columns',
+          options: {
+            columns: [
+              {
+                blocks: [
+                  {
+                    '@type': '@builder.io/sdk:Element',
+                    '@version': 2,
+                    id: 'builder-2f8d079276024173a80801d6fe0117f7',
+                    component: {
+                      name: 'Text',
+                      options: {
+                        text: 'third',
+                      },
+                    },
+                    responsiveStyles: {
+                      large: {
+                        display: 'flex',
+                        flexDirection: 'column',
+                        position: 'relative',
+                        flexShrink: '0',
+                        boxSizing: 'border-box',
+                        marginTop: '20px',
+                        lineHeight: 'normal',
+                        height: 'auto',
+                      },
+                    },
+                  },
+                ],
+              },
+              {
+                blocks: [
+                  {
+                    '@type': '@builder.io/sdk:Element',
+                    '@version': 2,
+                    id: 'builder-b6c8d6beff5b4049b6374827c59743fa',
+                    component: {
+                      name: 'Text',
+                      options: {
+                        text: 'second',
+                      },
+                    },
+                    responsiveStyles: {
+                      large: {
+                        display: 'flex',
+                        flexDirection: 'column',
+                        position: 'relative',
+                        flexShrink: '0',
+                        boxSizing: 'border-box',
+                        marginTop: '20px',
+                        lineHeight: 'normal',
+                        height: 'auto',
+                      },
+                    },
+                  },
+                ],
+              },
+            ],
+            space: 20,
+            stackColumnsAt: 'tablet',
+          },
+        },
+        responsiveStyles: {
+          large: {
+            position: 'relative',
+            flexShrink: '0',
+            boxSizing: 'border-box',
+            marginTop: '20px',
+            paddingBottom: '30px',
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'stretch',
+          },
+        },
+      },
+    ],
+  },
+};