+++ packages/sdks/src/components/block/block.lite.tsx
@@ -8,6 +8,7 @@ import {
   useStore,
   useTarget,
 } from '@builder.io/mitosis';
+import { TARGET } from '../../constants/target.js';
 import type {
   BuilderContextInterface,
   RegisteredComponents,
@@ -185,7 +186,7 @@ export default function Block(props: BlockProps) {
         registeredComponents: props.registeredComponents,
         builderBlock: state.processedBlock,
         includeBlockProps: state.blockComponent?.noWrap === true,
-        isInteractive: !state.blockComponent?.isRSC,
+        isInteractive: !(state.blockComponent?.isRSC && TARGET === 'rsc'),
       };
     },
   });
@@ -195,16 +196,21 @@ export default function Block(props: BlockProps) {
   });
 
   onMount(() => {
-    const blockId = state.processedBlock.id;
-    const animations = state.processedBlock.animations;
-    if (animations && blockId) {
-      bindAnimations(
-        animations.map((animation) => ({
-          ...animation,
-          elementId: blockId,
-        }))
-      );
-    }
+    useTarget({
+      reactNative: () => {},
+      default: () => {
+        const blockId = state.processedBlock.id;
+        const animations = state.processedBlock.animations;
+        if (animations && blockId) {
+          bindAnimations(
+            animations.map((animation) => ({
+              ...animation,
+              elementId: blockId,
+            }))
+          );
+        }
+      },
+    });
   });
 
   return (