+++ packages/core/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @builder.io/sdk
 
+## 3.0.1
+
+### Patch Changes
+
+- 69859d4: serialize functions for registering plugins so you can have showIf on fields as functions
+
+## 3.0.0
+
+### Major Changes
+
+- bc1d409: Fix: Reintroduced `JSON.stringify(userAttributes)` change to standardize parsing logic and preserve strings. This is a breaking change as it doesn't require manual stringification of `userAttributes` values. Ensure that attributes are not manually stringified before passing them to avoid potential issues.
+
+### Patch Changes
+
+- 1586519: Fix: remove `next: { revalidate: 1 }` in SDKs fetch
+
 ## 2.2.9
 
 ### Patch Changes