+++ packages/sdks/output/react-native/CHANGELOG.md
@@ -1,5 +1,43 @@
 # Builder.io React Native SDK Changelog (@builder.io/sdk-react-native)
 
+## 2.0.10
+
+### Patch Changes
+
+- 561e375: Fix: Drag and drop blocks into the correct place and add blocks using the "+Add Block" button
+
+## 2.0.9
+
+### Patch Changes
+
+- 69859d4: serialize functions for registering plugins so you can have showIf on fields as functions
+- f1527ee: Allow gap, rowGap, columnGap in react native strict mode styles
+
+## 2.0.8
+
+### Patch Changes
+
+- e8b80b3: Fix: scoped `isInteractive` prop for RSC SDK only so that it fixes Inner Layout > "Columns" option during visual editing
+
+## 2.0.7
+
+### Patch Changes
+
+- 6a4c1f5: Feature: add optional `strictStyleMode` prop to `<Content>` that will validate and ignore any style that does not work on React Native. The validation logic is identical to the one in the Visual Editor's "Strict React Native Styling" advanced setting.
+- 6a4c1f5: Fix: gracefully handle css sanitization errors.
+
+## 2.0.5
+
+### Patch Changes
+
+- 345086b: Fixes data bindings in Text blocks
+
+## 2.0.4
+
+### Patch Changes
+
+- 5a6cd39: Fix: ignore animations (avoid crashes when they are used).
+
 ## 2.0.3
 
 ### Patch Changes